Factored out some of the ansi/protocol.hpp header.
